% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine Two Sequences}
\usage{
combine(
  .f = function(x) (1:4)/4,
  .g = function(x) unlist(lapply(1:8, function(x) 2^x))
)
}
\arguments{
\item{.f}{A function returning a (periodic) sequence}

\item{.g}{A function returning a sequence}
}
\value{
An object of class \code{combine} containing:
\item{fg_dot}{The dot product of f and g: sum of \code{f[i]*g[i]}}
\item{fg_terms}{Element-wise products: \code{f[i]*g[i]} for each i}
\item{f_seq}{The periodic coefficient sequence}
\item{g_seq}{The base sequence}
}
\description{
This function combines two sequences. It is designed to explore series
with periodic coefficients.
}
\examples{
combine()

}
