% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_funcs.R
\name{complex_cyc}
\alias{complex_cyc}
\title{Complex Exponential Sequence}
\usage{
complex_cyc(inputs = seq(0, 2 * pi, length.out = 9)[-9], radians = TRUE)
}
\arguments{
\item{inputs}{A numeric input vector (in radians by default)}

\item{radians}{A Boolean. If TRUE (default), inputs are in radians; if FALSE, in degrees.}
}
\value{
A complex vector of values on the unit circle
}
\description{
Generates a complex exponential sequence for use as periodic coefficients in
series analysis. Returns e^(ix) = cos(x) + i*sin(x), enabling Fourier-like
complex analysis.The default excludes the endpoint to prevent duplicate values
when tiling.
}
\examples{
complex_cyc()
complex_cyc(seq(0, 2*pi, length.out = 17)[-17])
complex_cyc(c(0, 90, 180, 270), radians = FALSE)

}
