% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_funcs.R
\name{cos_cyc}
\alias{cos_cyc}
\title{Cosine Sequence}
\usage{
cos_cyc(inputs = seq(0, 2 * pi, length.out = 9)[-9], radians = TRUE)
}
\arguments{
\item{inputs}{A numeric input vector (in radians by default)}

\item{radians}{A Boolean. If TRUE (default), inputs are in radians; if FALSE, in degrees.}
}
\value{
A numeric vector of cosine values
}
\description{
Generates a cosine sequence for use as periodic coefficients in series analysis.
The default excludes the endpoint to prevent duplicate values when tiling.
}
\examples{
cos_cyc()
cos_cyc(seq(0, 2*pi, length.out = 17)[-17])
cos_cyc(c(0, 90, 180, 270), radians = FALSE)

}
