% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{plot.combine}
\alias{plot.combine}
\title{Plot Method for combine Objects}
\usage{
\method{plot}{combine}(x, legend_pos = "topleft", ...)
}
\arguments{
\item{x}{A combine object}

\item{legend_pos}{Position for the legend in plot 4 (default "topleft")}

\item{...}{Additional arguments (unused)}
}
\description{
Plot Method for combine Objects
}
