% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{print.combine}
\alias{print.combine}
\title{Print Method for combine Objects}
\usage{
\method{print}{combine}(x, n = 6, ...)
}
\arguments{
\item{x}{A combine object}

\item{n}{Number of terms to display (default 6)}

\item{...}{Additional arguments (unused)}
}
\description{
Print Method for combine Objects
}
