% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_funcs.R
\name{sin_cyc}
\alias{sin_cyc}
\title{Sine Sequence}
\usage{
sin_cyc(inputs = seq(0, 2 * pi, length.out = 9)[-9], radians = TRUE)
}
\arguments{
\item{inputs}{A numeric input vector (in radians by default)}

\item{radians}{A Boolean. If TRUE (default), inputs are in radians; if FALSE, in degrees.}
}
\value{
A numeric vector of sine values
}
\description{
Generates a sine sequence for use as periodic coefficients in series analysis.
The default excludes the endpoint to prevent duplicate values when tiling.
}
\examples{
sin_cyc()
sin_cyc(seq(0, 2*pi, length.out = 17)[-17])
sin_cyc(c(0, 90, 180, 270), radians = FALSE)

}
