% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_funcs.R
\name{square_cyc}
\alias{square_cyc}
\title{Square Wave Sequence}
\usage{
square_cyc(inputs = seq(0, 2 * pi, length.out = 9)[-9], radians = TRUE)
}
\arguments{
\item{inputs}{A numeric input vector (in radians by default)}

\item{radians}{A Boolean. If TRUE (default), inputs are in radians; if FALSE, in degrees.}
}
\value{
A numeric vector of square wave values (-1 or 1)
}
\description{
Generates a square wave sequence for use as periodic coefficients in series analysis.
The square wave alternates between -1 and 1, creating maximum contrast.
The default excludes the endpoint to prevent duplicate values when tiling.
}
\examples{
square_cyc()
square_cyc(seq(0, 2*pi, length.out = 17)[-17])
square_cyc(c(0, 90, 180, 270), radians = FALSE)

}
