% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{summary.combine}
\alias{summary.combine}
\title{Summary Method for combine Objects}
\usage{
\method{summary}{combine}(object, ...)
}
\arguments{
\item{object}{A combine object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary Method for combine Objects
}
