% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{combine}
\alias{combine}
\title{Combine Two Sequences}
\usage{
combine(
  .f = function(x) (1:4)/4,
  .g = function(x) unlist(lapply(1:8, function(x) 2^x))
)
}
\arguments{
\item{.f}{A function returning a (periodic) sequence}

\item{.g}{A function returning a sequence}
}
\value{
A list containing four elements:
\item{fg_dot}{The dot product of the .f and .g sequences}
\item{fg_terms}{The individual terms of the fg_dot series}
\item{f_seq}{The .f sequence}
\item{g_seq}{The .g sequence}
}
\description{
This function combines two sequences. It is designed to explore what series
with periodic coefficients
}
\examples{
combine()

}
