% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{plot.combine}
\alias{plot.combine}
\title{Plot Method for combine Objects}
\usage{
\method{plot}{combine}(x, legend_pos = "right", ...)
}
\arguments{
\item{x}{A combine object}

\item{legend_pos}{Position for legend. Either a keyword ("right", "topleft", etc.) or numeric coordinates.}

\item{...}{Additional arguments (unused)}
}
\description{
Plot Method for combine Objects
}
